package name change  com.ss.smartfilter to com.ss.smartfilterdemo
