+++ app/src/androidTest/java/com/ss/smartfilterdemo/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.ss.smartfilter
+package com.ss.smartfilterdemo
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.ss.smartfilter", appContext.packageName)
+        assertEquals("com.ss.smartfilterdemo", appContext.packageName)
     }
 }
