+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.SmartFilter"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.ss.smartfilterdemo.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.SmartFilter">
             <intent-filter>
