+++ app/src/main/java/com/ss/smartfilterdemo/MainActivity.kt
@@ -1,11 +1,11 @@
-package com.ss.smartfilter
+package com.ss.smartfilterdemo
 
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.material3.Surface
-import com.ss.smartfilter.screens.FilterScreen
-import com.ss.smartfilter.theme.smartFilterTheme
+import com.ss.smartfilterdemo.screens.FilterScreen
+import com.ss.smartfilterdemo.theme.smartFilterTheme
 
 
 class MainActivity : AppCompatActivity() {
