+++ app/src/main/java/com/ss/smartfilterdemo/screens/ExpandableSeatType.kt
@@ -1,4 +1,4 @@
-package com.ss.smartfilter.screens
+package com.ss.smartfilterdemo.screens
 
 import android.widget.LinearLayout
 import androidx.compose.foundation.layout.Column
@@ -25,7 +25,7 @@ import androidx.compose.ui.tooling.data.UiToolingDataApi
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.viewinterop.AndroidView
-import com.ss.smartfilter.addSingleSelectionListSeat
+import com.ss.smartfilterdemo.addSingleSelectionListSeat
 import com.ss.smartfilterlib.SmartFilter
 
 /**
